23.12.17 [app main f/w]  - warning 모두 제거 완료  - x-modem 동작 시나리오 수정 (app f/w 가동시에는 부트로더로 진입하도록 하고 사용자가 부트로더에 머물어 x-modem으로 f/w 다운로드 하도록 강제)

+++ CLI/CLI_TEST/boot/xmodem.c
@@ -130,7 +130,7 @@ BOOL_e XMODEM_GetRecord(U8 *u8DestAddress)
         u16CRC_check = (u16CRC_check<<8) ^ crc16tab[((u16CRC_check>>8) ^ u8GetChar)&0x00FF];
         
         #if 1
-        HAL_FLASH_Program(FLASH_TYPEPROGRAM_BYTE, u8DestAddress++, u8GetChar);
+        HAL_FLASH_Program(FLASH_TYPEPROGRAM_BYTE, (uint32_t)u8DestAddress++, (uint64_t)u8GetChar);
         #else
         *u8DestAddress++ = u8GetChar;
         #endif
